Begin to use resources instead of file paths
The applications come with several text and html files that they previously looked for in a hardcoded relative path, which made it hard to distribute just a jar file. I added those files as resources and then had to figure out how to adapt the api to use them. For the time being they are being written to temp files and then the file, instead of the path is being passed around. Eventually this will migrate to a InputStream system, but I wanted to make the change incremental at first.